Eliminate the Circuit
The `circuit` module was not really necessary as it only added an extra layer of abstraction in order to separate the logic of the actual simulation (`State`) and the execution and systematization of functions (`Circuit`).
This layer has been eliminated and its functionality has been added to the corresponding `State` classes.

The tests have been modified accordingly.
